+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fixed session plugin not receiving callbacks for nested archives containing only sub-archives, which caused sessions to never close and reports to not be generated
 - Delete rescan unused option and fix doc for period option.
 - Fixed report duration refers now to session duration
+- Analyzed volume is now based on input file when submit to gdetect
 
 ### Changed